v4.0

+++ option/src/app/java_helidon/rest/src/main/java/me/opc/mp/database/Dept.java
@@ -14,7 +14,7 @@ public class Dept {
 
     @Id
     @Column(name = "deptno", nullable = false, updatable = false)
-    private int deptno;
+    private long deptno;
 
     @Basic(optional = false)
     @Column(name = "dname")
@@ -27,7 +27,7 @@ public class Dept {
     public Dept() {
     }
 
-    public Dept( int deptno, String dname, String loc) {
+    public Dept( long deptno, String dname, String loc) {
         this.deptno = deptno;
         this.dname = dname;
         this.loc = loc;
@@ -37,7 +37,7 @@ public class Dept {
         return deptno;
     }
 
-    public void setDeptno(int deptno) {
+    public void setDeptno(long deptno) {
         this.deptno = deptno;
     }
