+++ option/src/app/java_micronaut/rest/src/main/java/com/example/Dept.java
@@ -7,7 +7,7 @@ import io.micronaut.serde.annotation.Serdeable;
 
 @Serdeable
 @MappedEntity
-public record Dept (@Id int deptno,
+public record Dept (@Id long deptno,
                     String dname,
                     String loc) {
 }
